# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ Vehicle.cpp
@@ -4,7 +4,7 @@
 #include "Vehicle.h"
 
 //constructor
-Vehicle::Vehicle(int cX, int cY, int end, double ang, double vel){
+Vehicle::Vehicle(int cX, int cY, int end, int ang, double vel){
     //values will be passed to GameState
     //values are placeholders for testing purposes
     velocity = vel;        //current velocity
@@ -14,4 +14,3 @@ Vehicle::Vehicle(int cX, int cY, int end, double ang, double vel){
     endurance = end;          //vehicle health
     //TODO: possibly include powerup classes / objects
 }
-//TODO: methods to handle movement, turning, speeding up, braking, etc.
